Add support to Ruby 3 kwargs dropping support to older Rails versions
The Rails shim / Action Pack integration method calls were failing with
Ruby 3 due to keyword args compatibility. In order to get apivore
working with Ruby 3, it was simpler to just drop older Rails version
support and fully move to the kwargs method.

This eliminates all the shim code for those integration method calls
(`get`, `post`, etc.), and also a few conditionals in the gemspec
related to older Ruby versions (since the required Ruby version in the
gemspec is already 2.6 anyway.)